requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
python_box==7.2.0
ruamel.base==1.0.0
